Create trajectory-point-properties.md
New general guidance section intended to capture the concerns raised in CR 134.  Table 6 Trajectory Point Properties was recreated to supply users with the information provided by the "Uses" and "Number of Instance Allowed" columns (which are not currently otherwise incorporated directly in the FIXM documentation). The additional guidance suggested in CR 134 for the HOLD_ENTRY, HOLD_EXIT, BEGIN_STAY, and END_STAY enumerations was added to the existing entry in the "Uses" column.