Update and rename non-airport-departures.md to non-aerodrome-departures.md
Changed "airport" to "aerodrome" to better match FIXM terminology.  Changed "and" to "or" in various places under FIXM Representation because current and previous departure information doesn't have to be of the same type (I believe you could file an update where current is departurePoint but previous was an aerodrome, or other variations).  Made estimatedOffBlockTime only be used for aerodrome departures because that's how I interpret the model documentation ("for use when a departure aerodrome does not exist ... or is otherwise unknown" - so, anytime you can't name the aerodrome, use route start time).  We may want to discuss this, though?  Other minor editorial changes.