Slent fixm 229 (#9)
* Create non-airport-departures.ms

* Delete non-airport-departures.ms

* Create non-airport-departures.md

* Update _sidebar.md

* Update _sidebar.md

* Update non-airport-departures.md

* Add files via upload

* Update non-airport-departures.md

* Update non-airport-departures.md

* Delete non-aerodrome-departures-01.png

* Add files via upload

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update geographical-positions.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update non-airport-departures.md

* Update _sidebar.md

Changed from "airport" to "aerodrome" to better match FIXM conventions.  Also, moved its position in the sidebar to be closer to similar sections.

* Update _sidebar.md

Minor cleanup.

* Update and rename non-airport-departures.md to non-aerodrome-departures.md

Changed "airport" to "aerodrome" to better match FIXM terminology.  Changed "and" to "or" in various places under FIXM Representation because current and previous departure information doesn't have to be of the same type (I believe you could file an update where current is departurePoint but previous was an aerodrome, or other variations).  Made estimatedOffBlockTime only be used for aerodrome departures because that's how I interpret the model documentation ("for use when a departure aerodrome does not exist ... or is otherwise unknown" - so, anytime you can't name the aerodrome, use route start time).  We may want to discuss this, though?  Other minor editorial changes.

---------

Co-authored-by: CPiroli-CTR <84804342+CPiroli-CTR@users.noreply.github.com>